generic IsErrNoRows in utils package
