+++ persistent/internal/driver/mgo/mgo_test.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/TykTechnologies/storage/persistent/databaseinfo"
+	"github.com/TykTechnologies/storage/persistent/utils"
 
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/mgo.v2"
@@ -178,14 +178,14 @@ func TestDelete(t *testing.T) {
 		// check if the object was deleted
 		err = driver.Query(ctx, object, &result, dbm.DBM{"_id": object.GetObjectID()})
 		assert.NotNil(t, err)
-		assert.True(t, driver.IsErrNoRows(err))
+		assert.True(t, utils.IsErrNoRows(err))
 	})
 
 	t.Run("deleting a non existent object", func(t *testing.T) {
 		// delete the object from the database
 		err := driver.Delete(ctx, object)
 		assert.NotNil(t, err)
-		assert.True(t, driver.IsErrNoRows(err))
+		assert.True(t, utils.IsErrNoRows(err))
 	})
 }
 
@@ -224,7 +224,7 @@ func TestUpdate(t *testing.T) {
 
 		err := driver.Update(ctx, object)
 		assert.NotNil(t, err)
-		assert.True(t, driver.IsErrNoRows(err))
+		assert.True(t, utils.IsErrNoRows(err))
 	})
 
 	t.Run("Updating an object without _id", func(t *testing.T) {
@@ -233,7 +233,7 @@ func TestUpdate(t *testing.T) {
 
 		err := driver.Update(ctx, object)
 		assert.NotNil(t, err)
-		assert.False(t, driver.IsErrNoRows(err))
+		assert.False(t, utils.IsErrNoRows(err))
 	})
 }
 
@@ -583,16 +583,6 @@ func TestUpdateAll(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestIsErrNoRows(t *testing.T) {
-	defer cleanDB(t)
-
-	mgoDriver := mgoDriver{}
-
-	assert.True(t, mgoDriver.IsErrNoRows(mgo.ErrNotFound))
-	assert.False(t, mgoDriver.IsErrNoRows(nil))
-	assert.False(t, mgoDriver.IsErrNoRows(mgo.ErrCursor))
 }
 
 func TestCount(t *testing.T) {
@@ -1937,5 +1927,5 @@ func TestGetDBType(t *testing.T) {
 	}
 
 	// ToDo: update for those cases where it returns aws
-	assert.Equal(t, databaseinfo.StandardMongo, info.Type)
+	assert.Equal(t, utils.StandardMongo, info.Type)
 }
