+++ persistent/internal/model/type.go
@@ -3,7 +3,7 @@ package model
 import (
 	"time"
 
-	"github.com/TykTechnologies/storage/persistent/databaseinfo"
+	"github.com/TykTechnologies/storage/persistent/utils"
 )
 
 type StorageLifecycle interface {
@@ -14,7 +14,7 @@ type StorageLifecycle interface {
 	Close() error
 
 	// DBType returns the type of the registered storage driver.
-	DBType() databaseinfo.DBType
+	DBType() utils.DBType
 }
 
 // DBTable is an interface that should be implemented by
